requirements.txt
discord.py==2.6.2
aiohttp==3.12.15
aiomysql==0.2.0
pydantic==2.11.7
APScheduler==3.11.0
loguru==0.7.3
dotenv==0.9.9
